requirements.txt
# Core requirements
streamlit==1.12.0
altair==4.2.0

# OpenAI and embeddings
openai>=1.12.0
tiktoken
faiss-cpu>=1.7.2

# Data processing
pandas>=1.5.0
numpy==1.23.0

# Environment/config
python-dotenv>=0.19.0
sentence-transformers>=2.2.2  # For local fallback

# Excel support
openpyxl>=3.0.0
xlrd>=2.0.0